CLN: MessageInput 컴포넌트 분리
